+++ qiniu/conf.js
@@ -11,15 +11,13 @@ var ua = function() {
 }
 
 exports.USER_AGENT = ua();
-
-exports.USER_AGENT = ua();
 exports.UP_HOST = 'http://upload.qiniu.com';
 exports.RS_HOST = 'http://rs.qbox.me';
 exports.RSF_HOST = 'http://rsf.qbox.me';
 exports.API_HOST = 'http://api.qiniu.com';
 exports.RPC_TIMEOUT = 3600000; // default rpc timeout: one hour
 exports.UC_HOST  = 'http://uc.qbox.me'; 
-exports.UP_HTTPS_HOST = 'https://up-z1.qbox.me';
-exports.SCHEME = 'http';
-exports.AUTOZONE = true;
-exports.DEADLINE = 0;
+exports.UP_HTTPS_HOST = 'https://up-z1.qbox.me'; 
+exports.SCHEME = 'http'; // 上传使用的协议 
+exports.AUTOZONE = true; // 自动获取不同机房空间的上传域名 
+exports.DEADLINE = 0; // 获取空间域名对应的缓存时间
