+++ qiniu/io.js
@@ -8,6 +8,7 @@ var mime = require('mime');
 var Readable = require('stream').Readable;
 var formstream = require('formstream');
 var urllib = require('urllib');
+var zone = require('./zone');
 
 exports.UNDEFINED_KEY = '?'
 exports.PutExtra = PutExtra;
@@ -33,98 +34,27 @@ function PutRet(hash, key) {
 }
 
 // onret: callback function instead of ret
-function putReadable(uptoken, key, rs, extra, onret) {
-  var scheme = conf.AUTOZONE;//上传域名使用的协议，默认使用http,可以在conf.js设置http 或 https
-  // 自动获取空间对应的上传域名
-  if(conf.AUTOZONE){
-    var ak = uptoken.toString().split(":")[0];
-    var tokenPolicy = uptoken.toString().split(":")[2];
-    var tokenPolicyStr = new Buffer(tokenPolicy, 'base64').toString();
-    var  json_tokenPolicyStr = JSON.parse(tokenPolicyStr);
-    var  bucket = json_tokenPolicyStr.scope;
-    // 判断过期时间
-    if( new Date().getTime() > conf.DEADLINE){
-      urllib.request('http://uc.qbox.me/v1/query', {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json'
-        },
-        data: {
-          'ak':ak,
-          'bucket':bucket
-        }
-      },function (err, data, res) {
-        var str = data.toString();
-        var json_str = JSON.parse(str);
-
-        //判断设置使用的协议
-        if(scheme == 'http'){
-           conf.UP_HOST = json_str.http.up[1];
-        }else{
-           conf.UP_HOST = json_str.https.up[0];
-        }
-
-      //ttl 获取up_host的缓存时间
-      conf.DEADLINE = json_str.ttl + new Date().getTime(); 
-
-     if (!extra) {
-        extra = new PutExtra();
-      }
-      if (!extra.mimeType) {
-        extra.mimeType = 'application/octet-stream';
-      }
-
-      if (!key) {
-        key = exports.UNDEFINED_KEY;
-      }
-
-      rs.on("error", function (err) {
-          onret({code: -1, error: err.toString()}, {});
-      });
-
-      var form = getMultipart(uptoken, key, rs, extra);
-      return rpc.postMultipart(conf.UP_HOST, form, onret);
-    });
-
-    }else{
-      if (!extra) {
-        extra = new PutExtra();
-      }
-      if (!extra.mimeType) {
-        extra.mimeType = 'application/octet-stream';
-      }
-
-      if (!key) {
-        key = exports.UNDEFINED_KEY;
-      }
-
-      rs.on("error", function (err) {
-          onret({code: -1, error: err.toString()}, {});
-      });
-
-      var form = getMultipart(uptoken, key, rs, extra);
-      return rpc.postMultipart(conf.UP_HOST, form, onret);
-    }
-  }else{
-    //指定空间上传域名，在conf.js中设置UP_HOST
-    if (!extra) {
-      extra = new PutExtra();
-    }
-    if (!extra.mimeType) {
-      extra.mimeType = 'application/octet-stream';
-    }
-
-    if (!key) {
-      key = exports.UNDEFINED_KEY;
-    }
-
-    rs.on("error", function (err) {
-        onret({code: -1, error: err.toString()}, {});
-    });
-
-    var form = getMultipart(uptoken, key, rs, extra);
-    return rpc.postMultipart(conf.UP_HOST, form, onret);
+function putReadable (uptoken, key, rs, extra, onret) {
+  if (!extra) {
+    extra = new PutExtra();
   }
+  if (!extra.mimeType) {
+    extra.mimeType = 'application/octet-stream';
+  }
+
+  if (!key) {
+    key = exports.UNDEFINED_KEY;
+  }
+
+  rs.on("error", function (err) {
+      onret({code: -1, error: err.toString()}, {});
+  });
+
+  var form = getMultipart(uptoken, key, rs, extra);
+  // 设置上传域名
+  zone.up_host(uptoken);
+
+  return rpc.postMultipart(conf.UP_HOST, form, onret);
 }
